NEWA ID updates, Jira users and a corner case
There are three main changes in this commit:

1. Previously we used issue action summary to look for existing
issues. This turned out to be not reliable because summary Jinja
templates can make it different for every respin. We decided to
drop summary in search and rather enode issues action ID into the
newa identifier. This works flaslessly but we had to re-factor
code related to newa identifier slightly and we had to associate
issues action to IssueHandler methods where it newa identifier
is used.

2. We found out that Jira have multiple ways how to identify user,
sometimes it is just kerberos login, other times it is either
email or e-mail with prefix. Fortunately, we can match the user
precisely using the email address. This also means we don't need
to extract user from the email in ErrataTool class anymore.

3. There is a specific corner case that can happen - a task got
dropped but its sub-tasks are not. Previously newa created new
task and re-used subtasks but they were still associated with the
old task and new one was without subtasks. This is now handled
by not reusing subtasks with dropped parent and replacing them
with new ones. This is the only solution possible since Jira REST
API (used in jira module) does not allow updates of parent field
in issues.