Reimplement recipe dimension merging and priorities
The former implementation was not consistent in which
definition has a priority. In general, CLI definition
should have a priority, however at the same time a user
must be able to override settings inherrited from the
previous NEWA command's YAML, like a 'compose'.
That was not possible due to a fact that CLI definitions
and YAML definitios have been merged prio building recipe
combinations.

This commit changes this behavior by separating CLI definition
and YAML definition. While YAML definition enters the recipe
configuration alghorithm as an initial_config which can be modified
by the recipe, values provided on CLI are provided separately
and the very end when building combinations and therefore have
a priority.

Such a change should enable user to define overrides on command line
while at the same time be able to adjust tmt/TF job configuration
in a recipe.