Implement recipe adjustments
Adds new section into a recipe file called 'adjustments'.
It enables to specify additinal configuration snippets that
would be applied on top of the configuration from 'fixtures'.
The primary use case is to apply configuration conditionally
using 'when' condition.
As well as 'fixtures', 'adjustments' can be imported from a file
or URL.